+++ src/algoritmia/_examples/backtracking/nqueens_bt.py
@@ -6,21 +6,15 @@ from algoritmia.schemes.bt_scheme import DecisionSequence, bt_solutions
 # Tipos  --------------------------------------------------------------------------
 
 type Decision = int  # Número de fila donde colocar la reina
-type Extra = None    # Este problema no necesita almacenar información adicional junto a la decisión
 
-# Queremos que una solución sea la secuencia de decisiones (números de fila) en forma de tupla:
+# Una solución es una secuencia de decisiones (números de fila) en forma de tupla:
 type Solution = tuple[Decision, ...]
-
-# - 'bt_solutions' y 'bt_vc_solutions' devuelven un Iterator con las DecisionSequence que
-#   llegan a una solución.
-# - Pero un objeto DecisionSequence no es una tupla de decisiones: debemos utilizar el método
-#   'decisions()' de la clase DecisionSequence para obtener la tupla.
 
 # --------------------------------------------------------------------------------
 
 
 def nqueens_solutions(board_size: int) -> Iterator[Solution]:
-    class NQueensDS(DecisionSequence[Decision, Extra]):  # Hemos definido Extra como None
+    class NQueensDS(DecisionSequence[Decision, None]):  # No necesitamos Extra -> ponemos None
         def is_solution(self) -> bool:
             return len(self) == board_size