+++ src/algoritmia/_examples/divide_and_conquer/convex_min.py
@@ -1,3 +1,5 @@
+# Mínimo de un vector convexo
+
 def convex_min(v: list[int]) -> int:
     def rec(start: int, end: int) -> int:
         if end - start == 1:
@@ -15,5 +17,5 @@ def convex_min(v: list[int]) -> int:
 
 
 if __name__ == '__main__':
-    v0 = [9, 8, 6, 5, 4, 4, 3, 2, 5, 6, 8]
+    v0 = [9, 8, 6, 5, 4, 4, 3, 2, 5, 6, 8]  # vector convexo
     print(convex_min(v0))