+++ src/algoritmia/_examples/divide_and_conquer/power.py
@@ -1,3 +1,5 @@
+# Potencia entera de un número
+
 def power_naif(a: float, n: int) -> float:
     if n == 0:
         return 1