+++ src/algoritmia/schemes/bab_scheme.py
@@ -1,5 +1,6 @@
 """
-Version: 6.0 (11-dic-2024)
+Version: 6.1 (12-oct-2025)
+         6.0 (11-dic-2024)
          5.2 (01-dic-2023)
          5.1 (23-nov-2023)
          5.0 (31-oct-2023)
@@ -7,7 +8,7 @@ Version: 6.0 (11-dic-2024)
          4.0 (23-oct-2021)
 
 @author: David Llorens (dllorens@uji.es)
-         (c) Universitat Jaume I 2023
+         (c) Universitat Jaume I 2025
 @license: GPL3
 """
 import operator
@@ -19,26 +20,37 @@ from algoritmia.datastructures.priorityqueues import MaxHeap, MinHeap, IPriority
 from algoritmia.schemes.bt_scheme import DecisionSequence, DecisionPath
 
 
-# ACERCA DEL TIPO ScoredSolution (importado de bt_scheme)
-# Las funciones bab_min_solve y bab_max_solve devuelven Optional[ScoredSolution]:
-#   - Devuelven 'None' si no hay solución.
-#   - Devuelven la tupla '(score, ds_sol)' si hay solución: donde 'ds_sol' es la
-#     BabDecisionSequence que lleva a la solución óptima y 'score', su puntuación.
+# --- Tipo Result[D, E, S] ---
 
-# ACERCA DEL CÁLCULO EFICIENTE DE LAS COTAS
+type Result[D, E, S] = tuple[S, BabDecisionSequence[D, E, S]] | None
+
+# Parámetros de tipo:
+#   - D: el tipo de una decisión
+#   - E: el tipo de la dataclass Extra
+#   - S: el tipo de la puntuación (int o float)
+
+# Un objeto de tipo Result[D, E, S] puede tomar los valores:
+#   - (score, ds_sol): si hay solución, donde 'ds_sol' es la BabDecisionSequence que lleva
+#     a la solución óptima y 'score', su puntuación.
+#   - None: si no hay solución.
+
+
+# --- Acerca del cálculo eficiente de las cotas ---
+
 # Dado que las cotas son inmutables solo deberían calcularse una vez:
 # - Los métodos calculate_opt_bound() y calculate_pes_bound() se llaman sólo en el constructor
 #   y sus resultados se guardan en los atributos self._opt y self._pes.
 # - Los métodos opt() y pes() sólo devuelven el valor de estos atributos.
 
+
 # La clase BabDecisionSequence -------------------------------------------------------
 
 
 @total_ordering  # Implementando < y ==, el resto de operadores de comparación se generan automáticamente
-class BabDecisionSequence[TDecision, TExtra, TScore](DecisionSequence[TDecision, TExtra]):
+class BabDecisionSequence[D, E, S](DecisionSequence[D, E]):
     def __init__(self,
-                 extra: TExtra | None = None,
-                 decisions: DecisionPath[TDecision] = (),
+                 extra: E | None = None,
+                 decisions: DecisionPath[D] = (),
                  length: int = 0):
         DecisionSequence.__init__(self, extra, decisions, length)
         self._pes = self.calculate_pes_bound()
@@ -47,11 +59,11 @@ class BabDecisionSequence[TDecision, TExtra, TScore](DecisionSequence[TDecision,
     # --- Métodos abstractos nuevos ---
 
     @abstractmethod
-    def calculate_opt_bound(self) -> TScore:  # Calcula y devuelve la cota optimista
+    def calculate_opt_bound(self) -> S:  # Calcula y devuelve la cota optimista
         pass
 
     @abstractmethod
-    def calculate_pes_bound(self) -> TScore:  # Calcula y devuelve la cota pesimista
+    def calculate_pes_bound(self) -> S:  # Calcula y devuelve la cota pesimista
         pass
 
     # --- Métodos abstractos heredados  ---
@@ -72,12 +84,12 @@ class BabDecisionSequence[TDecision, TExtra, TScore](DecisionSequence[TDecision,
 
     # Cota optimista. Para las soluciones su valor debe coincidir con la puntuación real
     @final
-    def opt(self) -> TScore:
+    def opt(self) -> S:
         return self._opt
 
     # Cota pesimista. Para las soluciones su valor debe coincidir con la puntuación real
     @final
-    def pes(self) -> TScore:
+    def pes(self) -> S:
         return self._pes
 
     # Comparar dos BabDecisionSequence es comparar sus cotas optimistas
@@ -92,23 +104,24 @@ class BabDecisionSequence[TDecision, TExtra, TScore](DecisionSequence[TDecision,
     # -- Métodos finales heredados que NO pueden sobreescribirse en las clases hijas ---
 
     # @final
-    # def add_decision(self, decision: TDecision, extra: TExtra = None) -> Self:    # O(1)
+    # def add_decision(self, decision: D, extra: E = None) -> Self:    # O(1)
 
     # @final
-    # def decisions(self) -> tuple[TDecision, ...]:     # O(n)
+    # def decisions(self) -> tuple[D, ...]:     # O(n)
 
     # @final
-    # def last_decision(self) -> TDecision:             # O(1)
+    # def last_decision(self) -> D:             # O(1)
 
     # @final
     # def __len__(self) -> int:                         # O(1)
 
 # Esquemas para BaB --------------------------------------------------------------------------
 
-def bab_solve[TDecision, TExtra, TScore](better: Callable[[TScore, TScore], bool],
-                                         heap: IPriorityQueue[TScore],
-                                         initial_ds: BabDecisionSequence[TDecision, TExtra, TScore]
-                                         ) -> tuple[TScore, BabDecisionSequence[TDecision, TExtra, TScore]] | None:
+
+def bab_solve[D, E, S](better: Callable[[S, S], bool],
+                       heap: IPriorityQueue[S],
+                       initial_ds: BabDecisionSequence[D, E, S]
+                      ) -> Result[D, E, S]:
     bps = initial_ds.pes()
     heap.add(initial_ds)
     best_seen = {initial_ds.state(): initial_ds.opt()}
@@ -128,11 +141,9 @@ def bab_solve[TDecision, TExtra, TScore](better: Callable[[TScore, TScore], bool
                     heap.add(new_ds)
 
 
-def bab_min_solve[TDecision, TExtra, TScore](initial_ds: BabDecisionSequence[TDecision, TExtra, TScore])\
-        -> tuple[TScore, BabDecisionSequence[TDecision, TExtra, TScore]] | None:
+def bab_min_solve[D, E, S](initial_ds: BabDecisionSequence[D, E, S]) -> Result[D, E, S]:
     return bab_solve(operator.lt, MinHeap(), initial_ds)
 
 
-def bab_max_solve[TDecision, TExtra, TScore](initial_ds: BabDecisionSequence[TDecision, TExtra, TScore])\
-        -> tuple[TScore, BabDecisionSequence[TDecision, TExtra, TScore]] | None:
+def bab_max_solve[D, E, S](initial_ds: BabDecisionSequence[D, E, S]) -> Result[D, E, S]:
     return bab_solve(operator.gt, MaxHeap(), initial_ds)